requirements.txt
streamlit==1.36.0
ultralytics
moviepy
opencv-python-headless
numpy
pillow
streamlit-drawable-canvas
